+++ nBodyCUDA/utils.h
@@ -1,4 +1,4 @@
 #pragma once
 #include "utils.h"
 //returns random number [0,1)
-float doubleRand();
+double doubleRand();
